# application/static/swagger.yaml
swagger: "2.0"
info:
  title: "Mechanic Shop API"
  description: "API for managing a mechanic shop - customers, service tickets, mechanics, and inventory"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: "Enter: Bearer <your_token>"
definitions:
...
  LoginCredentials:
    type: object
    required:
      - email
      - password
    properties:
      email:
        type: string
        example: "john@example.com"
      password:
        type: string
        example: "password123"

  # ---- RESPONSE DEFINITIONS ----
  LoginResponse:
    type: object
    properties:
      status:
        type: string
        example: "Success"
      message:
        type: string
        example: "successfully logged in."
      token:
        type: string
        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."

  # Customer definition
  # This is the payload for creating a new customer(information sent in the request body)
  CustomerPayload:
    type: object
    required:
      - name
      - email
      - phone
      - password
    properties:
      name:
        type: string
        example: "John Doe"
      email:
        type: string
        example: "john@example.com"
      phone:
        type: string
        example: "1234567890"
      password:
        type: string
        example: "password123"

  # Customer response definition
  # This is the response returned when a customer is created or retrieved from API
  CustomerResponse:
    type: object
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: "John Doe"
      email:
        type: string
        example: "john@example.com"
      phone:
        type: string
        example: "1234567890"

  # Customer list response definition
  # This is the response returned when a list of customers is retrieved from API
  CustomerListResponse:
    type: object
    properties:
      customers:
        type: array
        items:
          $ref: "#/definitions/CustomerResponse"
      total:
        type: integer
        example: 25
      pages:
        type: integer
        example: 3
      current_page:
        type: integer
        example: 1
  # Mechanic definition
...
  MechanicResponse:
    type: object
    properties:
      id:
        type: integer
        example: 1
      name:
        type: string
        example: "Jane Smith"
      email:
        type: string
        example: "jane@example.com"
      phone:
        type: string
        example: "0987654321"
      salary:
        type: string
        example: "50000"
  # Mechanic list response definition
...
  TicketResponse:
    type: object
    properties:
      id:
        type: integer
        example: 1
      service_date:
        type: string
        example: "2023-08-15"
      vin:
        type: string
        example: "1HGCM82633A123456"
      service_desc:
        type: string
        example: "Oil change and tire rotation"
      customer_id:
        type: integer
        example: 1
      mechanics:
        type: array
        items:
          $ref: "#/definitions/MechanicResponse"
      item:
        type: array
        items:
          $ref: "#/definitions/ItemResponse"
...
  TicketListResponse:
    type: array
    items:
      $ref: "#/definitions/TicketResponse"

  # Item definition
...
  ItemResponse:
    type: object
    properties:
      id:
        type: integer
        example: 1
      item_name:
        type: string
        example: "Brake Pads"
      price:
        type: number
        example: 45.99

  # Item list response definition
...
paths:
  /customers/login:
    post:
      tags:
        - "Customer"
      summary: "Customer Login, (generates token)"
      description: "Validates customer email and password, returns a JWT token to be used for protected routes."
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginResponse"
          examples:
            application/json:
              status: "Success"
              message: "successfully logged in."
              token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        400:
          description: "Invalid email or password"
  /customers/:
    post:
      tags:
        - "Customer"
      summary: "Create a customer"
      description: "Create a new customer account"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CustomerPayload"
      responses:
        201:
          description: "Customer created"
          schema:
            $ref: "#/definitions/CustomerResponse"
          examples:
            application/json:
              id: 1
              name: "John Doe"
              email: "john@example.com"
              phone: "1234567890"
        400:
          description: "Validation error or email already exists"
    get:
      tags:
        - "Customer"
      summary: "Get all customers (paginated)"
      description: "Retrieve paginated list of all customers"
      parameters:
        - in: query
          name: page
          type: integer
          description: "Page number (default: 1)"
        - in: query
          name: per_page
          type: integer
          description: "Items per page (default: 10)"
      responses:
        200:
          description: "List of customers"
          schema:
            $ref: "#/definitions/CustomerListResponse"
          examples:
            application/json:
              customers:
                - id: 1
                  name: "John Doe"
                  email: "john@example.com"
                  phone: "1234567890"
              total: 1
              pages: 1
              current_page: 1
    put:
      tags:
        - "Customer"
      summary: "Update the authenticated customer (requires token)"
      description: "Update the currently authenticated customer account using bearer token identity"
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CustomerPayload"
      responses:
        200:
          description: "Customer updated"
          schema:
            $ref: "#/definitions/CustomerResponse"
          examples:
            application/json:
              id: 1
              name: "John Doe Updated"
              email: "john@example.com"
              phone: "1234567890"
        400:
          description: "Validation error, missing/invalid token, or customer not found"
    delete:
      tags:
        - "Customer"
      summary: "Delete the authenticated customer (requires token)"
      description: "Delete the currently authenticated customer account using bearer token identity"
      security:
        - Bearer: []
      responses:
        200:
          description: "Customer deleted"
          examples:
            application/json:
              message: "Customer id: 1, Successfully deleted"
        400:
          description: "Missing/invalid token or customer not found"
  /customers/{id}:
    get:
      tags:
        - "Customer"
      summary: "Get a specific customer"
      description: "Retrieve customer details by ID"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: "Customer ID"
      responses:
        200:
          description: "Customer data"
          schema:
            $ref: "#/definitions/CustomerResponse"
          examples:
            application/json:
              id: 1
              name: "John Doe"
              email: "john@example.com"
              phone: "1234567890"
        400:
          description: "Customer not found"
  /customers/my-tickets:
    get:
      tags:
        - "Customer"
      summary: "Get my service tickets"
      description: "Retrieve all service tickets for authenticated customer - requires authentication"
      security:
        - Bearer: []
      responses:
        200:
          description: "List of service tickets"
          schema:
            $ref: "#/definitions/TicketListResponse"
          examples:
            application/json:
              - id: 10
                service_date: "2026-07-12"
                vin: "1HGCM82633A123456"
                service_desc: "Oil change"
                customer_id: 1
                mechanics: []
                item: []
        400:
          description: "Missing/invalid token"
...